requirements.txt
# Core LLM + LangChain packages
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.37
langchain-openai==0.1.7
langchain-huggingface==0.0.3
langchain-pinecone==0.0.3
langchain-experimental

# Vector DB - Pinecone
pinecone-client==3.0.0


# Embeddings
sentence-transformers==2.6.1
transformers>=4.40.0,<5.0.0
huggingface-hub>=0.16.4,<1.0

# PDF parsing
pypdf==3.17.4
pyreadline3
zstandard==0.22.0


# Web + Env
flask
python-dotenv

# -e .
